feat: Add `Level::None` for omitting the header
When combined with an empty `title`, this results in the header being
omitted entirely.

This isn't what @epage suggested in #167 since changing our rendering to
pass in a custom header is a bigger refactor than what I've been able to
get to. The goal of this change was to be very small without requiring
bigger code changes on our end. Unfortunately, this is a breaking
change, so I'm not sure what the appetite is for this.

Fixes #167